Extract dashboard widget controls
